requirements.txt
google-cloud-aiplatform
google-cloud-storage
google-genai
gitpython
google-adk
langchain-community
wikipedia
supabase
vnstock
transformers
crewai
crewai-tools